pull up LocationListRenderer to file-tree module
